Added per spawn costs to Powered Spawner Values to be decided.

+++ src/main/java/crazypants/enderio/machine/AbstractPoweredTaskEntity.java
@@ -139,7 +139,7 @@ public abstract class AbstractPoweredTaskEntity extends AbstractMachineEntity {
     return usePower(getPowerUsePerTick());
   }
 
-  protected double usePower(int wantToUse) {
+  protected int usePower(int wantToUse) {
     int used = Math.min(getEnergyStored(), wantToUse);
     setEnergyStored( Math.max(0, getEnergyStored() - used));
     currentTask.update(used);
